requirements.txt
# Core Framework
flask==2.3.3
werkzeug==2.3.7

# Data Processing
pandas==1.5.3
numpy==1.24.4
yfinance==0.2.28
pandas-ta==0.3.14b0

# HTTP Requests for Alpaca API
requests==2.31.0
alpaca-py==0.8.2

# Environment and Configuration
python-dotenv==1.0.0

# Caching and Rate Limiting
redis==5.0.1
Flask-Caching==2.1.0
Flask-Limiter==3.5.0

# Security
cryptography==41.0.7

# Database (Neon PostgreSQL)
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Additional requirements for production deployment
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3